Move models into separate files
Also.
